fix: warnings and errors under `#[cfg(test)]`

+++ src/stats/bhop.rs
@@ -74,7 +74,7 @@ impl fake::Dummy<fake::Faker> for BhopStats
 {
 	fn dummy_with_rng<R: rand::Rng + ?Sized>(_: &fake::Faker, rng: &mut R) -> Self
 	{
-		let total = rng.r#gen::<u16>();
+		let total = rng.r#gen::<u32>();
 		let perfs = rng.gen_range(0..=total);
 		let perfect_perfs = rng.gen_range(0..=perfs);
